+++ Kryupa/Core/Accounts/AccountView.swift
@@ -88,8 +88,19 @@ struct AccountView: View {
         case "About app":
             break
         case "Logout":
-            break
-           
+            
+            let primaryAction = UIAlertAction(title: "OK", style: .default) { action in
+                let domain = Bundle.main.bundleIdentifier!
+                UserDefaults.standard.removePersistentDomain(forName: domain)
+                UserDefaults.standard.synchronize()
+                NotificationCenter.default.post(name: .logout,
+                                                object: nil, userInfo: nil)
+            }
+            
+            let secondaryAction = UIAlertAction(title: "Cancel", style: .cancel)
+            
+            presentAlert(title: "Kryupa", subTitle: "Log Out",primaryAction: primaryAction,secondaryAction: secondaryAction)
+            
         default:
             break
         }
