+++ Kryupa/Core/Care Seeker Registration/Health Information Seeker/HealthInformationSeekerViewModel.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 import SwiftUI
-
+@MainActor
 class HealthInformationSeekerViewModel: ObservableObject{
     
     @Published var medicalConditionSelected: String = String()
